refactor: use CommonLib as submodule
