+++ .gitmodules
@@ -1,9 +1,9 @@
 [submodule "lib/spdlog"]
 	path = lib/spdlog
 	url = https://github.com/gabime/spdlog
-[submodule "proto"]
-	path = proto
-	url = https://github.com/SailGame/proto.git
 [submodule "lib/googletest"]
 	path = lib/googletest
 	url = https://github.com/google/googletest.git
+[submodule "lib/CommonLib"]
+	path = lib/CommonLib
+	url = https://github.com/SailGame/CommonLib.git
