+++ lib/CMakeLists.txt
@@ -1,2 +1,4 @@
 add_subdirectory(spdlog)
 add_subdirectory(googletest)
+set(BUILD_UNO_LIB ON)
+add_subdirectory(CommonLib)
